ref: BoradcastService to Broadcast::Post
Perks:
- Scalable for adding more Broadcast sub services
- More explicit
