+++ app/services/broadcast.rb
@@ -1,0 +1,5 @@
+module Broadcast
+  class << self
+  
+  end
+end
